Modif du Makefile

+++ Makefile
@@ -2,28 +2,33 @@
 CC= g++
 LIBNAME= Engine
 DIR= lib/
-LIBLINUX= -Iinclude -fPIC -lGL -lGLU -lGLEW -lSDL2 -lSDL2_image
-LIBWINDOWS= -Iinclude -I/c/MinGW/tiers/include -L/c/MinGW/tiers/lib -lopengl32 -lglu32 -lglew32 -lSDL2 -lSDL2_image
-FLAGS= -O3 -Wall -Werror -Wfatal-errors $(LIBLINUX)
+SYSTEM= $(shell uname)
+ifeq ($(SYSTEM), Linux)
+LIB= -Iinclude -fPIC -lGL -lGLU -lGLEW -lSDL2 -lSDL2_image
+else
+LIB= -Iinclude -I/c/MinGW/tiers/include -L/c/MinGW/tiers/lib -lopengl32 -lglu32 -lglew32 -lSDL2 -lSDL2_image
+endif
+FLAGS= -O3 -Wall -Werror -Wfatal-errors $(LIB)
 SRC= GL.cpp Object.cpp Window.cpp Camera.cpp FreeCam.cpp PlayerCam.cpp ShaderObject.cpp ShaderProgram.cpp GLcontext.cpp
 OBJ= $(SRC:%.cpp=obj/%.o)
 IMPORT= -Wl,--out-implib,$(DIR)lib$(LIBNAME).a
 
-all:	
-	@echo "make [os]"
-
-win:	$(OBJ) 
-	$(CC) -shared -o $(DIR)$(LIBNAME).dll $^ $(FLAGS) $(IMPORT)
-
-linux:	$(OBJ) 
+all:	$(OBJ)
+ifeq ($(SYSTEM), Linux)
 	$(CC) -shared -o $(DIR)lib$(LIBNAME).so $^ $(FLAGS)
+else
+	$(CC) -shared -o $(DIR)$(LIBNAME).dll $^ $(FLAGS) $(IMPORT)
+endif
 
 obj/%.o:	%.cpp
 	$(CC) -c -o $@ $^ $(FLAGS)
 
 clean:	
 	rm -f $(OBJ)
+ifeq ($(SYSTEM), Linux)
 	rm -f *~
 	rm -f *#
 	rm -f $(DIR)lib$(LIBNAME).so
+else
 	rm -f $(DIR)$(LIBNAME).dll $(DIR)lib$(LIBNAME).a
+endif
